Add exception for intent analysis
Issue-ID: USECASEUI-707

+++ intentanalysis/src/main/java/org/onap/usecaseui/intentanalysis/service/impl/ExpectationServiceImpl.java
@@ -19,6 +19,7 @@ package org.onap.usecaseui.intentanalysis.service.impl;
 
 import org.onap.usecaseui.intentanalysis.bean.models.Expectation;
 import org.onap.usecaseui.intentanalysis.bean.models.State;
+import org.onap.usecaseui.intentanalysis.exception.DataBaseException;
 import org.onap.usecaseui.intentanalysis.mapper.ExpectationMapper;
 import org.onap.usecaseui.intentanalysis.service.ExpectationService;
 import org.onap.usecaseui.intentanalysis.service.StateService;
@@ -26,6 +27,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.onap.usecaseui.intentanalysis.common.ResponseConsts;
 
 
 import java.util.ArrayList;
@@ -51,12 +53,22 @@ public class ExpectationServiceImpl implements ExpectationService {
                 stateService.createStateList(expectation.getStateList(), expectation.getExpectationId());
             }
         }
-        expectationMapper.insertExpectationList(expectationList, intentId);
+        int res = expectationMapper.insertExpectationList(expectationList, intentId);
+        if (res < 1) {
+            String msg = "Create expectation to database failed.";
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_INSERT_DATA_FAIL);
+        }
     }
 
     @Override
     public List<Expectation> getExpectationListByIntentId(String intentId) {
         List<Expectation> expectationList = expectationMapper.selectExpectationByIntentId(intentId);
+        if (expectationList == null) {
+            String msg = String.format("Intent id %s doesn't exist in database.", intentId);
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_QUERY_DATA_EMPTY);
+        }
         for (Expectation expectation : expectationList) {
             List<State> stateList = stateService.getStateListByExpectationId(expectation.getExpectationId());
             expectation.setStateList(stateList);
@@ -67,7 +79,17 @@ public class ExpectationServiceImpl implements ExpectationService {
     @Override
     public void deleteExpectationListById(String intentId) {
         List<Expectation> expectationList = expectationMapper.selectExpectationByIntentId(intentId);
-        expectationMapper.deleteExpectationByIntentId(intentId);
+        if (expectationList == null) {
+            String msg = String.format("Intent id %s doesn't exist in database.", intentId);
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_QUERY_DATA_EMPTY);
+        }
+        int res = expectationMapper.deleteExpectationByIntentId(intentId);
+        if (res < 1) {
+            String msg = "Delete expectation in database failed.";
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_DELETE_DATA_FAIL);
+        }
         for (Expectation expectation : expectationList) {
             stateService.deleteStateListByExpectationId(expectation.getExpectationId());
         }
@@ -77,8 +99,9 @@ public class ExpectationServiceImpl implements ExpectationService {
     public void updateExpectationListById(List<Expectation> expectationList, String intentId) {
         List<Expectation> expectationDBList = expectationMapper.selectExpectationByIntentId(intentId);
         if (expectationDBList == null) {
-            LOGGER.error("Intent ID {} doesn't exist in database.", intentId);
-            throw new IllegalArgumentException("This intent ID doesn't exist in database.");
+            String msg = String.format("Intent id %s doesn't exist in database.", intentId);
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_QUERY_DATA_EMPTY);
         }
         List<String> expectationDBIdList = new ArrayList<>();
         for (Expectation expectationDB : expectationDBList) {
@@ -88,7 +111,12 @@ public class ExpectationServiceImpl implements ExpectationService {
         for (Expectation expectation : expectationList) {
             if (expectationDBIdList.contains(expectation.getExpectationId())) {
                 stateService.updateStateListByExpectationId(expectation.getStateList(), expectation.getExpectationId());
-                expectationMapper.updateExpectation(expectation);
+                int res = expectationMapper.updateExpectation(expectation);
+                if (res < 1) {
+                    String msg = "Update expectation in database failed.";
+                    LOGGER.error(msg);
+                    throw new DataBaseException(msg, ResponseConsts.RET_UPDATE_DATA_FAIL);
+                }
                 expectationDBIdList.remove(expectation.getExpectationId());
             } else {
                 expectationService.insertExpectation(expectation, intentId);
@@ -102,11 +130,21 @@ public class ExpectationServiceImpl implements ExpectationService {
 
     @Override
     public void insertExpectation(Expectation expectation, String intentId) {
-        expectationMapper.insertExpectation(expectation, intentId);
+        int res = expectationMapper.insertExpectation(expectation, intentId);
+        if (res < 1) {
+            String msg = "Create expectation to database failed.";
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_INSERT_DATA_FAIL);
+        }
     }
 
     @Override
     public void deleteExpectationById(String expectationId) {
-        expectationMapper.deleteExpectationById(expectationId);
+        int res = expectationMapper.deleteExpectationById(expectationId);
+        if (res < 1) {
+            String msg = "Delete expectation in database failed.";
+            LOGGER.error(msg);
+            throw new DataBaseException(msg, ResponseConsts.RET_DELETE_DATA_FAIL);
+        }
     }
 }
